+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.12.0] - 2023-04-18
+### Changed
+- Changed the behavior of `ImsHelper.recipe_upload()` to only skip uploading a
+  recipe if the artifact MD5 sums match, not just if the IMS image names match.
+- Change `ImsHelper.image_upload_artifacts()` to check equality of images
+  based on artifact checksums in addition to image names.
+
 ## [2.11.1] - 2023-02-06
 ### Fixed
 - Fix bug where an `"ims_image_record"` key was not being returned in